Rebuild the DOM to a file
Parser, working with DomTree, can now recreate html to a file. It is similar, but not the same as original. For one, it's too complicated to add indentation and keep text all on one line when it has embedded tags. So, every parsed tag and text element has its own line.